A bunch fixes for warnings in the build.

+++ src/condor_util_lib/email.c
@@ -73,8 +73,8 @@ email_open( const char *email_addr, const char *subject )
 
 	/* Take care of the subject. */
 	if ( subject ) {
-		int prolog_length = strlen(EMAIL_SUBJECT_PROLOG);
-		int subject_length = strlen(subject);
+		size_t prolog_length = strlen(EMAIL_SUBJECT_PROLOG);
+		size_t subject_length = strlen(subject);
 		FinalSubject = malloc(prolog_length + subject_length + 1);
 		memcpy(FinalSubject, EMAIL_SUBJECT_PROLOG, prolog_length);
 		memcpy(&FinalSubject[prolog_length], subject, subject_length);
